+++ .pylintrc
@@ -12,6 +12,7 @@ disable = all,
 enable = anomalous-backslash-in-string,
          bad-super-call,
          bare-except,
+         cyclic-import,
          dangerous-default-value,
          function-redefined,
          line-too-long,
