+++ nengo/builder/operator.py
@@ -36,7 +36,7 @@ import numpy as np
 
 import nengo.utils.numpy as npext
 from nengo.utils.compat import is_array_like
-from nengo.utils.connection import function_name
+from nengo.utils.functions import function_name
 from nengo.exceptions import BuildError, SimulationError
