+++ nengo/utils/functions.py
@@ -9,6 +9,25 @@ from nengo.exceptions import ValidationError
 from nengo.utils.compat import is_number, iteritems
 
 
+def function_name(func):
+    """Returns the name of a function.
+
+    Unlike accesing ``func.__name__``, this function is robust to the
+    different types of objects that can be considered a function in Nengo.
+
+    Parameters
+    ----------
+    func : callable or array_like
+        Object used as function argument.
+
+    Returns
+    -------
+    str
+        Name of function object.
+    """
+    return getattr(func, "__name__", func.__class__.__name__)
+
+
 def piecewise(data):
     """Create a piecewise constant function from a dictionary.
